yupik: dts: Nuke oplus,dsi-brightness-remapping.
This mapping is used to map a logarithmic brightness
curve to an linear one. This leads to effects in AOSP,
where the brightness slider has visible effect starting
by around 50%.

If the dts element isn't existent, then the brightness
is set in a linear way, which gives much better results.

Change-Id: Ie557da8ce4e59491c2bdbd1239f429403e4e2e8a